services/question-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
motor==3.3.2
pydantic==2.5.3
python-dotenv==1.0.0
redis==5.0.1
apscheduler==3.10.4
scikit-learn==1.4.0
numpy==1.26.3

